requirements.txt
# Django et Django REST Framework
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentification et sécurité
djangorestframework-simplejwt==5.3.1
django-filter==23.5
cryptography==42.0.0

# Documentation API
drf-spectacular==0.27.0

pymysql

# Environnement
python-decouple==3.8
python-dotenv==1.0.0

# Utilitaires
Pillow
pytz==2023.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
